# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OrbiCheck API
  description: API for validation, deduplication, and risk assessment services
  version: 1.0.0
  contact:
    name: OrbiCheck Team
    email: support@orbicheck.com
servers:
  - url: https://api.orbicheck.com/v1
    description: Production server
  - url: https://dev-api.orbicheck.com/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /validate/email:
    post:
      tags:
        - Runtime API
      summary: Validate email address
      description: Performs comprehensive email validation including format, MX records, and disposable detection
      operationId: validateEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  description: The email address to validate
      responses:
        '200':
          description: Email validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    description: Whether the email is valid
                  normalized:
                    type: string
                    description: Normalized email address
                  disposable:
                    type: boolean
                    description: Whether the email is from a disposable service
                  mx_found:
                    type: boolean
                    description: Whether MX records were found
                  reason_codes:
                    type: array
                    items:
                      type: string
                    description: List of validation reason codes
                  request_id:
                    type: string
                  ttl_seconds:
                    type: integer
                    description: Cache TTL in seconds
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Error message
        request_id:
          type: string
          description: Request identifier
...